% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{check_na}
\alias{check_na}
\title{check_na}
\usage{
check_na(col_name, data)
}
\arguments{
\item{col_name}{A string representing the name of the column in \code{data}}

\item{data}{A dataframe containing the dataset.}
}
\value{
A warning specifying the columns with missing values
}
\description{
This function checks for NAs in a specified column and
returns a warning if there are any
}
