% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_time_interval}
\alias{get_time_interval}
\title{get_time_interval}
\usage{
get_time_interval(data = NULL, time = NULL, group = NULL)
}
\arguments{
\item{data}{an object of class 'data.frame' containing equally spaced
time series data (daily, weekly, monthly data) for one or multiple spatial units or groups.}

\item{time}{A string representing the name of the column in \code{data} that contains time information.
The column should be in "yyyy-mm-dd" format.}

\item{group}{(Optional) A string representing the names of the columns in \code{data}
that contains grouping information. Required if \code{data} contains multiple spatial units.}
}
\value{
A character string indicating the time interval of the dataset. Possible values:
"day", "week", "month", "year", or "unknown".
}
\description{
This function determines the time interval (e.g., daily, weekly, monthly, yearly)
of a dataset based on the differences between the first two time points.
}
