% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHRexplore-package.R
\docType{package}
\name{GHRexplore-package}
\alias{GHRexplore}
\alias{GHRexplore-package}
\title{GHRexplore: Exploratory Analysis of Temporal and Spatio-Temporal Health Data}
\description{
A collection of commonly used visualizations of temporal and spatio-temporal health data including case counts, incidence rates, and covariates. The available plot types include time series, heatmaps, seasonality plots, maps and more. The package supports standard data transformations such as temporal and spatial aggregations, while offering extensive customization options for the resulting figures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.earth.bsc.es/ghr/ghrexplore}
  \item \url{https://bsc-es.github.io/GHRtools/docs/GHRexplore/GHRexplore}
  \item Report bugs at \url{https://gitlab.earth.bsc.es/ghr/ghrexplore/-/issues}
}

}
\author{
\strong{Maintainer}: Carles Milà \email{carles.milagarcia@bsc.es} (\href{https://orcid.org/0000-0003-0470-0760}{ORCID})

Authors:
\itemize{
  \item Giovenale Moirano (\href{https://orcid.org/0000-0001-8748-3321}{ORCID})
  \item Anna B. Kawiecki (\href{https://orcid.org/0000-0002-0499-2612}{ORCID})
  \item Rachel Lowe (\href{https://orcid.org/0000-0003-3939-7343}{ORCID}) [copyright holder]
}

}
\keyword{internal}
