% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_MS.R
\docType{data}
\name{map_MS}
\alias{map_MS}
\title{Administrative Map for Municipalities in the Mato Grosso do Sul}
\format{
\subsection{}{

A simple feature (sf) object including 11 rows and 2 columns:
\describe{
\item{code}{Unique ID number to each
micro region (11 units)}
\item{geometry}{geometries of the sf multipolygon}
}
}
}
\usage{
map_MS
}
\description{
A map including 11 municipalities.
}
\keyword{datasets}
