% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_common.R
\name{plot_multiple_common}
\alias{plot_multiple_common}
\title{plot_multiple_common}
\usage{
plot_multiple_common(
  plot_function,
  data,
  var,
  type = "cov",
  var_label = NULL,
  pop = NULL,
  pt = 1e+05,
  palette = "IDExtremes",
  ...
)
}
\arguments{
\item{plot_function}{indicates which of the ghr plot types to use
Options are: 'plot_timeseries', 'plot_heatmap', 'plot_seasonality', 'plot}

\item{data}{an object of class "data.frame" containing covariates or cases}

\item{var}{String specifying the name of the variable of interest.}

\item{type}{character that specifies the type of variable in \code{var}.
Possible values include 'cov' (covariate, default), 'counts' (case counts),
and 'inc' (case incidence). If \code{type='inc'}, \code{pop} is required.}

\item{var_label}{character vector with custom names for the case or
covariate variables.
If cases are to be plotted, the label for the case variable must be the first
in the vector. This vector must be the same length as the number of
covariates and cases specified  in the 'cases' and 'var' arguments}

\item{pop}{only needed if "cases" is passed as argument. the name of the
variable that identifies the population.
Necessary if type = "incidence".}

\item{pt}{only needed if "cases" is passed as argument. the scale of the
person-time (default 100,000) for incidence rates.
Necessary if type = "incidence".}

\item{palette}{when character vector with length of the number of variables
to be plotted, custom palettes for each plot; when single character, same
palette is applied to all plots.
Options are: "GreenPurp","Purp","Green","BlueRed","Blue","Red",
"WaterSand","Water","Sand","Qualitative", "Sunset","Earth","BlYlRd","RdGnBl"}

\item{...}{Additional arguments passed to \code{plot_function}.}
}
\value{
a list of plots, each list element named after the variable plotted
}
\description{
outputs multiple plots, each representing one variable
}
\details{
variable names and palette can be customized for each plot, otherwise
parameters will be the same for each variable plotted (options depend on the
plot type chosen)
}
\examples{
# plots <- plot_multiple(
# plot_function = plot_timeseries,
# data = dengue_MS,
# var = c("tmax", "tmin", "pdsi"),
# cases = "dengue_cases",
# pop = "population",
# type = "incidence",
# var_label = c("Dengue Cases", "Max Temp", "Min Temp", "PDSI"),
# palettes = c("Blue", "Red", "Earth", "Sunset"),
# time = "date",
# area = "micro_code")
# # Access individual plots
# print(plots[["dengue_cases"]])  # Dengue cases plot
# print(plots[["tmax"]])          # Max Temp plot
# print(plots[["tmin"]])          # Min Temp plot
# print(plots[["pdsi"]])      # PDSI plot
}
