% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_map.R
\name{plot_multiple_map}
\alias{plot_multiple_map}
\title{plot_multiple_map}
\usage{
plot_multiple_map(
  plot_function,
  data,
  var,
  type,
  var_label = NULL,
  pop = NULL,
  pt = 1e+05,
  palette = "IDExtremes",
  ...
)
}
\arguments{
\item{plot_function}{indicates which of the ghr plot types to use
Options are: 'plot_map'}

\item{data}{an object of class "data.frame" containing covariates or cases}

\item{var}{character vector the name of the variables identifying the
variables to be plotted.}

\item{type}{character that specifies the type of variable in \code{var}.
Possible values include 'cov' (covariate, default), 'counts' (case counts),
and 'inc' (case incidence). If \code{type='inc'}, \code{pop} is required.}

\item{var_label}{character vector with custom names. This vector must be the
same length as the number of covariates and cases specified  in 'var' arguments.}

\item{pop}{only needed if "cases" is passed as argument. the name of the
variable that identifies the population.
Necessary if type = "incidence".}

\item{pt}{the scale of the person-time (default 100,000) for incidence rates.
only needed if type = "incidence".}

\item{palette}{when character vector with length of the number of variables
to be plotted, custom palettes for each plot; when single character, same
palette is applied to all plots.
Options are: "GreenPurp","Purp","Green","BlueRed","Blue","Red",
"WaterSand","Water","Sand","Qualitative", "Sunset","Earth","BlYlRd","RdGnBl"}

\item{...}{Additional arguments passed to \code{plot_function}.}
}
\value{
a list of plots, each list element named after the variable plotted
}
\description{
outputs multiple choropleth map plots, each representing one variable
}
\details{
variable names and palette can be customized for each plot, otherwise
parameters will be the same for each variable plotted
}
\examples{
# plots <- plot_multiple(
# plot_function = plot_timeseries,
# data = dengue_MS,
# var = c("tmax", "tmin", "pdsi"),
# cases = "dengue_cases",
# pop = "population",
# type = "incidence",
# var_label = c("Dengue Cases", "Max Temp", "Min Temp", "PDSI"),
# palettes = list("Blue", "Red", "Earth", "Sunset"),
# map = map_MS,       
# map_area = "code",  
# time = "date",
# area = "micro_code")
# # Access individual plots
# print(plots[["dengue_cases"]])  # Dengue cases plot
# print(plots[["tmax"]])          # Max Temp plot
# print(plots[["tmin"]])          # Min Temp plot
# print(plots[["pdsi"]])      # PDSI plot

}
