% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{ymd_strict}
\alias{ymd_strict}
\title{ymd_strict}
\usage{
ymd_strict(x)
}
\arguments{
\item{x}{a string to be parsed}
}
\value{
an aggregated dataset
}
\description{
Checks parsing of a date accoring to yyyy-mm-dd
}
