% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{aggregate_cases}
\alias{aggregate_cases}
\title{aggregate_cases}
\usage{
aggregate_cases(
  data = NULL,
  cases = NULL,
  pop = NULL,
  time = NULL,
  area = NULL,
  pt = 1e+05,
  aggregate_space = NULL,
  aggregate_time = NULL
)
}
\arguments{
\item{data}{an object of class "data.frame" containing equally spaced
(daily, weekly, monthly ) incident cases for one or multiple locations.}

\item{cases}{the name of variable that identify cases}

\item{pop}{the name of the variable that identify the population.
Necessary if type = "incidence".}

\item{time}{the name of the variable that identify the temporal dimension
of the data.frame. The values must be in date format ("dd-mm-yyyy")
representing the day of observation if the data.frame has a
daily observations, the first day of the week if the data.frame has weekly
observations, or the first day of the month if the data.frame has a monthly
observations.}

\item{area}{the name of the variable identifying the finest spatial unit
if the data contains observations form multiple areas}

\item{pt}{the scale of the person-time (default 100000) for incidence rates.}

\item{aggregate_space}{the name of the variable over which perform the
spatial aggregation.}

\item{aggregate_time}{one of the three option among "week", "month", "year".
Specify the temporal scale over which perform the aggregation.}
}
\value{
an aggregated dataset
}
\description{
aggregate_cases
}
