% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{aggregate_cov}
\alias{aggregate_cov}
\title{aggregate_cov}
\usage{
aggregate_cov(
  data = NULL,
  var = NULL,
  time = NULL,
  area = NULL,
  aggregate_space = NULL,
  aggregate_time = NULL,
  aggregate_space_fun = "mean",
  aggregate_time_fun = "mean"
)
}
\arguments{
\item{data}{an object of class "data.frame" containing equally spaced
(daily, weekly, monthly ) incident cases for one or multiple locations.}

\item{var}{the name of variable that identify the covariate}

\item{time}{the name of the variable that identify the temporal dimension
of the data.frame. The values must be in date format ("dd-mm-yyyy")
representing the day of observation if the data.frame has a
daily observations, the first day of the week if the data.frame has weekly
observations, or the first day of the month if the data.frame has a monthly
observations.}

\item{area}{the name of the variable identifying the finest spatial unit
if the data contains observations form multiple areas}

\item{aggregate_space}{the name of the variable over which perform the
spatial aggregation.}

\item{aggregate_time}{one of the three option among "week", "month", "year".
Specify the temporal scale over which perform the aggregation.}

\item{aggregate_space_fun}{character indicating the function to be performed
in the aggregation over space, default is "mean"}

\item{aggregate_time_fun}{character indicating the function to be performed
in the aggregation over time, default is "mean"}
}
\value{
an aggregated dataset
}
\description{
aggregate_cov
}
