% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dengue_SP.R
\docType{data}
\name{dengue_SP}
\alias{dengue_SP}
\title{Dengue cases in \emph{Sao Paulo}}
\format{
A data frame with 678 rows and 8 columns:
\describe{
\item{date}{First day of the week in date format.}
\item{geocode}{Unique ID code for Sao Paulo microregion.}
\item{cases}{Number of notified dengue cases.}
\item{year}{Year (2010 - 2022).}
\item{temp_med}{Weekly average daily mean temperature.}
\item{precip_tot}{Weekly cumulative precipitation.}
\item{enso}{El Niño-Southern Oscillation index.}
\item{pop}{Number of inhabitants.}
}
}
\source{
\url{https://info.dengue.mat.br/services/api}
}
\usage{
dengue_SP
}
\description{
Data frame containing the weekly number of notified dengue cases in
the municipality of \emph{Sao Paulo}, as well as a set of climatic covariates.
}
\keyword{datasets}
