% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GHR_palette.R
\name{GHR_palette}
\alias{GHR_palette}
\alias{GHR_colors}
\alias{GHR_palettes}
\title{GHR_palette()
Generate GHR Color Palettes}
\usage{
GHR_palette(palette = "IDExtremes")

GHR_colors(n, palette = "IDExtremes")

GHR_palettes()
}
\arguments{
\item{palette}{string, name of the palette}

\item{n}{numerical, number of colours}
}
\value{
GHR_palette() returns the function that generates the color palette and the
attribute 'na_color'.\cr
GHR_colors() returns a vector of the colors.\cr
GHR_palettes() returns a plot with the custom GHR palettes.

a plot with all available palettes
}
\description{
Generates friendly color palettes with color ranges useful in
GHRexplore plotting including custom, ColorBrewer and viridis palettes.
See all available options by running \code{GHR_palettes()},
\code{RColorBrewer::display.brewer.all()} and visiting \verb{https://search.r-project.org/CRAN/refmans/viridisLite/html/viridis.html}.
}
\examples{
GHR_palettes()
}
