% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_MS.R
\docType{data}
\name{map_MS}
\alias{map_MS}
\title{Municipality boundaries of \emph{Mato Grosso do Sul}}
\format{
A simple feature (sf) object including 11 rows and 2 columns:
\describe{
\item{code}{Unique ID number to each micro region (11 units).}
\item{geometry}{Geometries of the sf multipolygon.}
}
}
\usage{
data(map_MS)
}
\description{
Administrative boundaries (polygon geometries) of the 11 municipalities
in \emph{Mato Grosso do Sul} (Brazil) to be used with the \code{dengue_MS} data set.
}
\keyword{datasets}
