% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{plot_correlation}
\usage{
plot_correlation(
  data = NULL,
  var = NULL,
  method = "pearson",
  title = NULL,
  palette = "IDExtremes",
  var_label = NULL,
  print = FALSE
)
}
\arguments{
\item{data}{an object of class "data.frame" containing spatio(-temporal)
covariates.}

\item{var}{a character vector containing variables for which the correlation
matrix will be calculated}

\item{method}{a method between "pearson" or "spearman" or "kendall, default "pearson"}

\item{title}{title of the plot.}

\item{palette}{GHR, RColorBrewer or viridisLite palette. See all available
options by running \code{GHR_palettes()}, \code{RColorBrewer::display.brewer.all()}
and visiting \verb{https://search.r-project.org/CRAN/refmans/viridisLite/html/viridis.html}}

\item{var_label}{a character vector containing custom names for the variables}

\item{print}{boolean. If TRUE, print the correlation matrix.}
}
\value{
a correlation matrix and heatmap
}
\description{
plot_correlation
}
\examples{
# Load data
data("dengue_MS")

# Pearson correlation heatmap
plot_correlation(dengue_MS, 
                 method = "pearson",
                 var = c("dengue_cases","pop_density", 
                         "tmax", "tmin", "pdsi", "urban",
                         "water_network", "water_shortage"),  
                 var_label = c("dengue cases","pop. density", 
                               "max temp", "min temp", "drought index", "urbanization",
                               "water network", "water shortage"),
                 title = "Correlation matrix",
                 palette = "magma") 

# Print spearman correlation matrix
plot_correlation(dengue_MS,
                 method = "spearman",
                 var = c("dengue_cases","pop_density", 
                         "tmax", "tmin", "pdsi", "urban",
                         "water_network", "water_shortage"),  
                 var_label = c("dengue cases","pop. density", 
                               "max temp", "min temp", "drought index", "urbanization",
                               "water network", "water shortage"),
                 print = TRUE) 

}
