% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(
  data,
  var = NULL,
  type = "cov",
  time = NULL,
  pop = NULL,
  pt = 1e+05,
  area = NULL,
  aggregate_space = NULL,
  aggregate_time = "month",
  aggregate_space_fun = "mean",
  aggregate_time_fun = "mean",
  log = FALSE,
  title = NULL,
  var_label = NULL,
  ylab = NULL,
  xlab = NULL,
  palette = "IDExtremes",
  centering = NULL
)
}
\arguments{
\item{data}{an object of class "data.frame" containing equally spaced
(daily, weekly, monthly) observations for one or multiple locations.}

\item{var}{the name of the column identifying the variable of interest.}

\item{type}{character that specifies the type of variable in \code{var}.
Possible values include 'cov' (covariate, default), 'counts' (case counts),
and 'inc' (case incidence). If \code{type='inc'}, \code{pop} is required.}

\item{time}{the name of the variable that identifies the temporal dimension
of the data.frame. The values must be in date format ("yyyy-mm-dd")
representing the day of observation for daily data, the first day of the
week for weekly, or the first day of the month for monthly observations.}

\item{pop}{character identifying the variable name for population. only needed
if \code{type='inc'}.}

\item{pt}{numerical only used for \code{type='inc'}. It represents the scale of the
person-time (default 100,000) for incidence rates.}

\item{area}{the name of the variable that identifies the different locations
(e.g., areal units). It is used when multiple time-series plots are required.}

\item{aggregate_space}{the name of the variable over which to perform
spatial aggregation.}

\item{aggregate_time}{specifies the temporal scale over which to perform
temporal aggregation. Options are: "week", "month".}

\item{aggregate_space_fun}{character indicating a function to be performed
in the aggregation over space, default is "mean". Only for covariates.
For cases the only function available is the mean}

\item{aggregate_time_fun}{character indicating a function to be performed
in the aggregation over space, default is "mean". Only for covariates.
For cases the only function available is the mean}

\item{log}{logical, default FALSE. Log-transform the variable of interest.}

\item{title}{title of the plot.
Default is "'var'" or "'var_label'" if var_label is specified}

\item{var_label}{character with a custom name for the case or covariate variable}

\item{ylab}{Label for the y-axis}

\item{xlab}{Label for the x-axis}

\item{palette}{GHR, RColorBrewer or viridisLite palette. See all available
options by running \code{GHR_palettes()}, \code{RColorBrewer::display.brewer.all()}
and visiting \verb{https://search.r-project.org/CRAN/refmans/viridisLite/html/viridis.html}}

\item{centering}{numerical or "median", defaults to NULL. If set,
it centers the palette on that value.}
}
\value{
A heatmap for a given variable.
}
\description{
Plots a matrix of covariates or cases (case counts or case incidence).
}
\details{
Time of the observation (month, week, or day) is represented on the
x-axis, year is represented on the y-axis.
}
\examples{

# Load data
data("dengue_MS")

# Covariate heatmap with space aggregation
plot_heatmap(dengue_MS,
             var = "tmin",
             var_label = "Minimum temp.",
             type = "cov",
             time = "date",
             area = "micro_code",
             aggregate_space = "meso_code",  
             aggregate_space_fun = "mean",   
             palette = "BrBG")

# Case count heatmap with log scale with palette centering
plot_heatmap(dengue_MS,
             var = "dengue_cases", 
             type = "counts",
             time = "date",  
             area = "micro_code",  
             palette = "Reds", 
             title = "Dengue counts", 
             var_label = "Dengue \ncounts",
             log = TRUE)  
             
# Case incidence (for 1,000 persons) heatmap with space aggregation
plot_heatmap(dengue_MS,
             var = "dengue_cases", # Variable defining the cases 
             type = "inc",
             pop = "population",
             pt = 1000,
             time = "date",          
             area = "micro_code", 
             aggregate_space = "meso_code", 
             palette = "Purp")            
}
