% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_map}
\alias{plot_map}
\title{plot_map}
\usage{
plot_map(
  data,
  var = NULL,
  type = "cov",
  time = NULL,
  pop = NULL,
  pt = 1e+05,
  area = NULL,
  map = NULL,
  map_area = NULL,
  by_year = TRUE,
  aggregate_time_fun = "mean",
  log = FALSE,
  title = NULL,
  var_label = NULL,
  palette = "IDExtremes",
  centering = NULL,
  bin = NULL
)
}
\arguments{
\item{data}{an object of class "data.frame" containing equally spaced
(daily, weekly, monthly) observations for one or multiple locations.}

\item{var}{the name of the column identifying the variable of interest.}

\item{type}{character that specifies the type of variable in \code{var}.
Possible values include 'cov' (covariate, default), 'counts' (case counts),
and 'inc' (case incidence). If \code{type='inc'}, \code{pop} is required.}

\item{time}{the name of the variable that identifies the temporal dimension
of the data.frame. The values must be in date format ("yyyy-mm-dd")
representing the day of observation for daily data, the first day of the
week for weekly, or the first day of the month for monthly observations.}

\item{pop}{the name of the variable that identifies the population.}

\item{pt}{the scale of the person-time (default 100,000) for incidence rates.}

\item{area}{the name of the variable that identifies the different locations
(e.g., areal units).}

\item{map}{the name of the sf object corresponding to the spatial unit
specified in 'area'.}

\item{map_area}{the name of the variable that identifies the different locations
(e.g., areal units) in the map object. If not specified, it assumes the same
name as 'area'.}

\item{by_year}{logical, if TRUE a map for each year is produced.}

\item{aggregate_time_fun}{character indicating the function to be performed in
the temporal aggregation ("mean","median"), default is "mean".}

\item{log}{logical, default FALSE. Log-transform the variable of interest.}

\item{title}{title of the plot.
Default is "'var'" or "'var_label'" if var_label is specified}

\item{var_label}{character with a custom name for the case or covariate variable}

\item{palette}{GHR, RColorBrewer or viridisLite palette. See all available
options by running \code{GHR_palettes()}, \code{RColorBrewer::display.brewer.all()}
and visiting \verb{https://search.r-project.org/CRAN/refmans/viridisLite/html/viridis.html}}

\item{centering}{numerical or "median", defaults to NULL. If set,
it centers the palette on that value.}

\item{bin}{a vector of percentiles to bin the variable of interest, e.g.
\code{seq(0, 1, 0.25)} for quartiles. Defaults to NULL (no binning).}
}
\value{
a map representing the incidence rates.
}
\description{
Plots a choropleth map of case incidence
}
\details{
Longitude is represented on the x-axis, latitude on the y axis,
and the case incidence in each spatial unit is represented by a distinct color.
The option to plot case counts without the population denominator is not given,
as case counts on a map would be a representation of the underlying population
}
\examples{
# Load data
library("sf")
data("dengue_MS")
data("map_MS")

# Temporal average of a covariate
plot_map(data = dengue_MS, 
         var = "tmin",    
         type = "cov",
         time = "date",       
         area = "micro_code",  
         map = map_MS,         
         map_area = "code",   
         aggregate_time_fun = "mean",         
         palette ="Reds",
         by_year = FALSE,
         var_label= "Min Temp.")

# Categorical covariate
plot_map(data = dengue_MS, 
         var = "biome_name",        
         time = "date",      
         area = "micro_code", 
         map = map_MS,       
         map_area = "code",  
         palette ="viridis",
         by_year = FALSE,
         var_label= "Biome")

# Case counts by year
dengue_MS \%>\% 
  plot_map(var = "dengue_cases",         # Variable to be plotted 
           type = "counts",
           time = "date",        # Variable defining the date "dd-mm-yyyy"
           area = "micro_code",  # Variable defining area in the dataframe
           pop = "population",
           map = map_MS,         # the sf object for the map 
           map_area = "code",
           palette = "Reds")

# Case incidence (for 1,000 persons) by year
plot_map(dengue_MS,
         var = "dengue_cases",         # Variable to be plotted 
         type = "inc",
         time = "date",        # Variable defining the date "dd-mm-yyyy"
         area = "micro_code",  # Variable defining area in the dataframe
         pop = "population",
         pt = 1000,
         map = map_MS,         # the sf object for the map 
         map_area = "code",   
         palette = "viridis")
}
